backend/requirements.txt
Django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.8
django-cors-headers>=3.7
Pillow>=8.0
